requirements.txt
amqp==5.3.1
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
bidict==0.23.1
billiard==4.2.2
blinker==1.9.0
cachetools==6.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
cycler==0.12.1
docutils==0.22.2
fastapi==0.120.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.59.2
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.9
kombu==5.5.4
MarkupSafe==3.0.3
matplotlib==3.10.6
MouseInfo==0.1.3
mss==10.1.0
netifaces==0.10.6
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pydantic==2.12.3
pydantic_core==2.41.4
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==2.0.1
PyMySQL==1.1.2
pyparsing==3.2.4
pyperclip==1.9.0
PyRect==0.2.0
pyscreenrec==0.6
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.14.2
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
redis==5.2.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==3.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
wsproto==1.2.0
